backend/requirements.txt
numpy>=1.21.0
tensorflow>=2.8.0
pillow>=9.0.0
Flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0
